Split biosero lighthouse service elements out into separate file and added exception test
