+++ modules/__mocks__/lighthouse_service_biosero.js
@@ -1,0 +1,4 @@
+// https://jestjs.io/docs/en/jest-object#jestcreatemockfrommodulemodulename
+const lighthouseBiosero = jest.createMockFromModule('@/modules/lighthouse_service_biosero').default
+
+export default lighthouseBiosero
