+++ pages/biosero_cherrypick.vue
@@ -43,6 +43,7 @@
 
 import BioseroCherrypickForm from '@/components/BioseroCherrypickForm'
 import lighthouse from '@/modules/lighthouse_service'
+import lighthouse_biosero from '@/modules/lighthouse_service_biosero'
 import Alert from '@/components/Alert'
 
 export default {
@@ -75,11 +76,11 @@ export default {
       this.getDataFromLighthouse(lighthouse.getFailureTypes(), 'failureTypes')
     },
     async create(form) {
-      const response = await lighthouse.createDestinationPlateBiosero(form)
+      const response = await lighthouse_biosero.createDestinationPlateBiosero(form)
       this.handleResponse(response)
     },
     async fail(form) {
-      const response = await lighthouse.failDestinationPlateBiosero(form)
+      const response = await lighthouse_biosero.failDestinationPlateBiosero(form)
       this.handleResponse(response)
     },
     handleResponse(response) {
